+++ rio-backend/src/config/mod.rs
@@ -1,13 +1,17 @@
 pub mod bindings;
 pub mod colors;
 pub mod defaults;
+pub mod keyboard;
 pub mod navigation;
+pub mod renderer;
 pub mod theme;
 pub mod window;
 
 use crate::config::bindings::Bindings;
 use crate::config::defaults::*;
+use crate::config::keyboard::Keyboard;
 use crate::config::navigation::Navigation;
+use crate::config::renderer::Renderer;
 use crate::config::window::Window;
 use colors::Colors;
 use log::warn;
@@ -22,26 +26,6 @@ pub enum ConfigError {
     ErrLoadingConfig(String),
     ErrLoadingTheme(String),
     PathNotFound,
-}
-
-#[derive(Default, Debug, Serialize, Deserialize, PartialEq, Clone, Copy)]
-pub enum Performance {
-    #[default]
-    High,
-    Low,
-}
-
-impl std::fmt::Display for Performance {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Performance::High => {
-                write!(f, "High")
-            }
-            Performance::Low => {
-                write!(f, "Low")
-            }
-        }
-    }
 }
 
 #[derive(Default, Debug, Serialize, Deserialize, PartialEq, Clone)]
@@ -90,16 +74,12 @@ pub struct Config {
     pub navigation: Navigation,
     #[serde(default = "Window::default")]
     pub window: Window,
-    #[serde(default = "Performance::default")]
-    pub performance: Performance,
     #[serde(default = "default_shell")]
     pub shell: Shell,
-    #[serde(default = "bool::default", rename = "disable-unfocused-render")]
-    pub disable_unfocused_render: bool,
     #[serde(default = "default_use_fork", rename = "use-fork")]
     pub use_fork: bool,
-    #[serde(default = "bool::default", rename = "use-kitty-keyboard-protocol")]
-    pub use_kitty_keyboard_protocol: bool,
+    #[serde(default = "Keyboard::default")]
+    pub keyboard: Keyboard,
     #[serde(default = "default_working_dir", rename = "working-dir")]
     pub working_dir: Option<String>,
     #[serde(rename = "line-height", default = "default_line_height")]
@@ -141,8 +121,10 @@ pub struct Config {
     pub ignore_selection_fg_color: bool,
     #[serde(default = "default_bool_true", rename = "confirm-before-quit")]
     pub confirm_before_quit: bool,
-    #[serde(default = "default_bool_true", rename = "hide-cursor-when-typing")]
+    #[serde(default = "bool::default", rename = "hide-cursor-when-typing")]
     pub hide_cursor_when_typing: bool,
+    #[serde(default = "Renderer::default")]
+    pub renderer: Renderer,
 }
 
 #[cfg(not(target_os = "windows"))]
@@ -370,16 +352,15 @@ impl Default for Config {
             colors: Colors::default(),
             cursor: default_cursor(),
             scroll: Scroll::default(),
-            use_kitty_keyboard_protocol: false,
+            keyboard: Keyboard::default(),
             developer: Developer::default(),
-            disable_unfocused_render: false,
             env_vars: default_env_vars(),
             fonts: SugarloafFonts::default(),
             line_height: default_line_height(),
             navigation: Navigation::default(),
             option_as_alt: default_option_as_alt(),
             padding_x: default_padding_x(),
-            performance: Performance::default(),
+            renderer: Renderer::default(),
             shell: default_shell(),
             theme: String::default(),
             use_fork: default_use_fork(),
@@ -443,9 +424,12 @@ mod tests {
         "#,
         );
 
-        assert!(!result.disable_unfocused_render);
+        assert!(!result.renderer.disable_unfocused_render);
 
-        assert_eq!(result.performance, Performance::default());
+        assert_eq!(
+            result.renderer.performance,
+            renderer::Performance::default()
+        );
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
 
@@ -461,14 +445,17 @@ mod tests {
     fn test_if_explict_defaults_match() {
         let result = create_temporary_config("defaults", &default_config_file_content());
 
-        assert_eq!(result.performance, Performance::default());
+        assert_eq!(
+            result.renderer.performance,
+            renderer::Performance::default()
+        );
         assert_eq!(result.env_vars, default_env_vars());
         assert_eq!(result.cursor, default_cursor());
         assert_eq!(result.theme, String::default());
         assert_eq!(result.cursor, default_cursor());
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.shell, default_shell());
-        assert!(!result.disable_unfocused_render);
+        assert!(!result.renderer.disable_unfocused_render);
         assert_eq!(result.use_fork, default_use_fork());
         assert_eq!(result.line_height, default_line_height());
 
@@ -495,7 +482,10 @@ mod tests {
 
         let result = Config::load_from_path(&file_name);
 
-        assert_eq!(result.performance, Performance::default());
+        assert_eq!(
+            result.renderer.performance,
+            renderer::Performance::default()
+        );
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
         // Colors
@@ -506,15 +496,18 @@ mod tests {
     }
 
     #[test]
-    fn test_change_config_performance() {
+    fn test_change_config_renderer() {
         let result = create_temporary_config(
             "change-performance",
             r#"
+            [renderer]
             performance = "Low"
+            backend = "Vulkan"
         "#,
         );
 
-        assert_eq!(result.performance, Performance::Low);
+        assert_eq!(result.renderer.performance, renderer::Performance::Low);
+        assert_eq!(result.renderer.backend, renderer::Backend::Vulkan);
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
         // Colors
@@ -533,7 +526,7 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::High);
+        assert_eq!(result.renderer.performance, renderer::Performance::High);
         assert_eq!(result.env_vars, [String::from("A=5"), String::from("B=8")]);
         assert_eq!(result.cursor, default_cursor());
         assert_eq!(result.fonts, SugarloafFonts::default());
@@ -562,7 +555,8 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::High);
+        assert_eq!(result.renderer.performance, renderer::Performance::High);
+        assert_eq!(result.renderer.backend, renderer::Backend::Automatic);
         assert_eq!(result.cursor, '_');
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
@@ -582,7 +576,7 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::High);
+        assert_eq!(result.renderer.performance, renderer::Performance::High);
         assert_eq!(result.option_as_alt, String::from("Both"));
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
@@ -603,7 +597,10 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::default());
+        assert_eq!(
+            result.renderer.performance,
+            renderer::Performance::default()
+        );
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
         // Colors
@@ -625,7 +622,10 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::default());
+        assert_eq!(
+            result.renderer.performance,
+            renderer::Performance::default()
+        );
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, String::default());
         // Bindings
@@ -640,10 +640,12 @@ mod tests {
         let result = create_temporary_config(
             "change-style",
             r#"
-            performance = "Low"
             font-size = 14.0
             line-height = 2.0
             padding-x = 0.0
+
+            [renderer]
+            performance = "Low"
 
             [window]
             background-opacity = 0.5
@@ -656,7 +658,7 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::Low);
+        assert_eq!(result.renderer.performance, renderer::Performance::Low);
         assert_eq!(result.fonts.size, 14.0);
         assert_eq!(result.line_height, 2.0);
         assert_eq!(result.padding_x, 0.0);
@@ -685,7 +687,7 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::High);
+        assert_eq!(result.renderer.performance, renderer::Performance::High);
         assert_eq!(result.fonts, SugarloafFonts::default());
         assert_eq!(result.theme, "lucario");
         // Colors
@@ -836,15 +838,17 @@ mod tests {
         let result = create_temporary_config(
             "change-use-fork",
             r#"
-            performance = "Low"
-            disable-unfocused-render = true
             use-fork = true
+
+            [renderer]
+            disable-unfocused-render = true
+            performance = "Low"
         "#,
         );
 
-        assert_eq!(result.performance, Performance::Low);
+        assert_eq!(result.renderer.performance, renderer::Performance::Low);
         // Advanced
-        assert!(result.disable_unfocused_render);
+        assert!(result.renderer.disable_unfocused_render);
         assert!(result.use_fork);
 
         // Colors
@@ -868,11 +872,13 @@ mod tests {
     }
 
     #[test]
-    fn test_change_developer() {
+    fn test_change_developer_and_performance() {
         let result = create_temporary_config(
             "change-developer",
             r#"
+            [renderer]
             performance = "Low"
+            backend = "GL"
 
             [developer]
             enable-fps-counter = true
@@ -880,7 +886,8 @@ mod tests {
         "#,
         );
 
-        assert_eq!(result.performance, Performance::Low);
+        assert_eq!(result.renderer.performance, renderer::Performance::Low);
+        assert_eq!(result.renderer.backend, renderer::Backend::GL);
         // Developer
         assert_eq!(result.developer.log_level, String::from("INFO"));
         assert!(result.developer.enable_fps_counter);