+++ sugarloaf/examples/term.rs
@@ -61,7 +61,7 @@ async fn main() {
 
     let mut sugarloaf = Sugarloaf::new(
         &sugarloaf_window,
-        wgpu::PowerPreference::HighPerformance,
+        sugarloaf::SugarloafRenderer::default(),
         sugarloaf::font::fonts::SugarloafFonts::default(),
         // "Fira Code".to_string(),
         // "Monaco".to_string(),